mainRouteDetail fails to parse Bird output
In my setup, the code parsing the Bird output didn't work. To be honest, I am a little surprised by the fix, so please review and let me know what you think.

Anyway, I studied the regular expressions and figured that the properties come out of the matches array in the order shown in my patch. Previous order was yielding broken results, with fields in incorrect places, etc.

I also tried to cover the case of unreachable, unicast or blackhole routes (if branch in line 151), but cannot extensively test these.